requirements.txt
ark-nlp==0.0.9
datasets==2.19.1
huggingface-hub==0.23.2
lightning==2.2.5
nltk==3.8.1
opencv-python==4.9.0.80
pandas==2.2.2
Pebble==5.0.7
peft==0.11.1
pillow==10.2.0
PyMuPDF==1.24.5
python-Levenshtein==0.25.1
pytorch-forecasting==1.0.0
pytorch-lightning==1.5.9
scikit-learn==1.5.0
seqeval==1.2.2
tensorboard==2.16.2
timm==0.6.12
torchmetrics==0.11.4
tqdm==4.66.4
transformers==4.41.2
ujson==5.10.0
pypandoc_binary
pypandoc